% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-manual.R
\name{bin_manual}
\alias{bin_manual}
\title{Bin the variables of a dataset manually}
\usage{
bin_manual(.data, bad, ..., check = FALSE)
}
\arguments{
\item{.data}{A tibble or data.frame}

\item{bad}{A variable name indicating which rows are bad borrowers}

\item{...}{Named lists separated by commas containing mappings of variables
to a transformation. Cuts for each variable are separated by \code{c_l} or
\code{c_r}.}

\item{check}{A logical, if TRUE, \code{bin_manual} will check for a minimum
of 30 borrowers for each bin. FALSE by default.}
}
\value{
A tibble or data.frame with the variables transformed into character
  vectors.
}
\description{
\code{bin_manual} cuts numeric, variables according to the mapping specified
by the user. This is a standard step in feature engineering or data
transformation for scorecard modelling.
}
\details{
Similar to the dplyr \code{mutate}, the function takes a named list of lists,
separated by commas, with the variable name on the left side, and the
transformation on the right side. \code{bin_manual} uses the \code{c_l} and
\code{c_r} functions to close where the values of a variable will be cut and
transformed into categories.
}
\examples{
bin_manual(german, bad, age = c_l(40, 55))
bin_manual(german,
           bad,
           duration = c_r(15, 32),
           check = TRUE)
}
