% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed.R
\name{c_l}
\alias{c_l}
\alias{c_r}
\title{Define value cuts for binning}
\usage{
c_l(...)

c_r(...)
}
\arguments{
\item{...}{Numbers, separated by commas indicating interval cuts in ascending
order}
}
\value{
A list of lists containing each cut and a length (1) character vector
    containing the symbol for the bounds.
}
\description{
Functions for explicitly defining value cuts and how the intervals are closed
either left or right.
}
\examples{
c_l(23, 50, 60)
c_r(2.5, 4, 5)
}
