% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit-score.R
\name{credit_score}
\alias{credit_score}
\title{Score input data using the scorecard model}
\usage{
credit_score(.data, model)
}
\arguments{
\item{.data}{A tibble of borrower information with the variable names that
match the variable names of the scorecard model}

\item{model}{A scorecard model object}
}
\value{
A numeric vector of credit scores, whether scaled or not
}
\description{
Borrower data is placed into \code{credit_score} together with the model to
score each borrower.
}
