% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-logit.R
\name{fit_logit}
\alias{fit_logit}
\title{Fit data into a logistic regression model}
\usage{
fit_logit(.data, formula)
}
\arguments{
\item{.data}{A tibble or data.frame}

\item{formula}{A formula of the form \code{y ~ x_1 + x_2 + ...}, where
\code{y} is the bad variable and \code{x1_2}, \code{x_2}, \code{...} are
the variables that will be used in the scorecard.}
}
\value{
A glm object containing the resulting model, with the name of the
    bad variable, and the weights of evidence for each value in each
    variable as attributes.
}
\description{
\code{fit_logit} takes a dataset and formula input by the user to generate
a logistic regression model that will be later used for the scoring borrower
accounts.
}
\details{
The function performs the following:
\enumerate{
    \item Calculate the weights-of-evidence (WOEs) of the categorical
        variables
    \item Replace the category values with the weights of evidence
    \item Returns a glm model object with modified attributes
}

All numeric variables are used in the logistic regression model as is. All
categorical variables are converted into WOEs before modelling. For better
accuracy and for simplicity, it is recommended that numeric variables are
first \link[=bin_manual]{binned} before modelling.
}
\examples{
library(dplyr)

bin_manual(german, bad, age = c_l(40, 55)) \%>\%
  fit_logit(bad ~ age + history + purpose)

bin_manual(german,
           bad,
           duration = c_r(15, 32),
           check = TRUE) \%>\%
 fit_logit(bad ~ duration + age + employed_since)
#

}
