% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-double.R
\name{scale_double_odds}
\alias{scale_double_odds}
\title{Scale model output to a format using points to double and default points at
odds of 50-to-1}
\usage{
scale_double_odds(score_model, odds_fifty = 600, pdo = 20)
}
\arguments{
\item{score_model}{A modified model object containing the model and
associated WOEs}

\item{odds_fifty}{A numeric, the point score of a borrower with a probability
of default of 50 to 1, default input is 600 points}

\item{pdo}{A numeric, score points given doubling odds of a default, defaults
to 20 points}
}
\value{
A modified model object containg the offset and factor computed
    by the function
}
\description{
\code{scale_double_odds} takes a scorecard model and formats the output using
points-to-double and the set score representing odds of default at 50-to-1.
Both are input by the user.
}
\details{
This is useful for representing the credit scores in the three-digit format
common in credit scoring companies like Equifax, FICO, and TransUnion.
Scaling is performed not to improve model predictability, but to improve
readability by stakeholders. This will generally depend on where the credit
scores are being used.
}
