% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-manual.R
\name{scale_manual}
\alias{scale_manual}
\title{Scale model output to a format using manual input}
\usage{
scale_manual(score_model, offset = 500, factor = 20)
}
\arguments{
\item{score_model}{A modified model object containing the model and
associated WOEs}

\item{offset}{A numeric, the credit score given default}

\item{factor}{A numeric, multiply the probability of default with this
factor}
}
\value{
A modified model object containg the offset and factor computed
    by the function
}
\description{
\code{scale_manual} takes a scorecard model and formats the output using
a manual offset and factor input
}
\details{
This is useful for representing the credit scores in the three-digit format
common in credit scoring companies like Equifax, FICO, and TransUnion.
Scaling is performed not to improve model predictability, but to improve
readability by stakeholders. This will generally depend on where the credit
scores are being used.
}
